online-inference/custom-basnet/client/requirements.txt
requests==2.23.0
Pillow==7.1.2
